Moved History commands to its own module.
